Rearrange utility functions
